+++ KansyaTeamGame2/ObjMagicalGirl.h
@@ -20,11 +20,12 @@ class CObjMagicalGirl : public CObj
 		float m_gx;//位置
 		float m_gy;
 		float m_postrue;//姿勢
-		float m_mp;//MP
-		float m_maxmp;//MAXMP
-		float m_atk_animation;//攻撃のモーション
-		float h_maxhp;//主人公の最大HP
-		float h_hp;//主人公の現HP
+
+		int m_mp;//MP
+		int m_maxmp;//MAXMP
+		int m_atk_animation;//攻撃のモーション
+		int h_maxhp;//主人公の最大HP
+		int h_hp;//主人公の現HP
 		
 		int m_mtime;//MP回復用時間
 		int m_btime;//ホーミング弾用時間
